Added registry for loading schemas
